feat: add getCards funtion
