+++ src/components/CardList/CardListItem/index.tsx
@@ -1,12 +1,13 @@
 import React, { FC } from 'react';
 import { ICardListItem } from '../../../interfaces/searchCard';
+import vite from '../../../../public/vite.svg';
 
 const CardListItem: FC<ICardListItem> = ({ card }) => {
   return (
     <div data-testid="card" className="card" key={card.id}>
       <div className="card__image">
         <img
-          src={`https://image.tmdb.org/t/p/w300/${card.poster_path}`}
+          src={card.poster_path ? `https://image.tmdb.org/t/p/w300/${card.poster_path}` : vite}
           alt=""
           className="card__img"
         />
